added code for chatbot frontend
