--- web-frontend/flaskIntro/.github/ISSUE_TEMPLATE.md
@@ -1,9 +1,0 @@
-Thanks for checking out my Flask tutorial repo! Since this is a tutorial repository, it's not updated unless a major breaking issue is discovered.
-Because of the reason mentioned, please follow the checklist below when opening an issue.
-
-Make sure to add **all the information needed to understand the problem** so that I can help. If the info is missing I'll add the 'Needs more information' label and close the issue until there is enough information.
-
-- [ ] Provide a **minimal code snippet** that reproduces the bug.
-- [ ] Provide **screenshots** where appropriate
-- [ ] What's the **version** of Python you're using?
-- [ ] Are you using Mac, Linux or Windows?
