+++ web-frontend/test/templates/index.html
@@ -1,0 +1,184 @@
+<!-- <!DOCTYPE html>
+<html>
+
+<head>
+    <title>Chatbot</title>
+</head>
+
+<body>
+    <h1>Chatbot</h1>
+    <form action="/" method="POST">
+        <input type="text" name="user_input">
+        <input type="submit" value="Send">
+    </form>
+    {% if response %}
+    <p>{{ response }}</p>
+    {% endif %}
+</body>
+
+</html> -->
+<!DOCTYPE html>
+<html>
+
+<head>
+    <meta charset="UTF-8">
+    <title>ChatGPT</title>
+    <style>
+        body {
+            background-color: #f2f2f2;
+            font-family: Arial, sans-serif;
+        }
+
+        .chatbot-container {
+            display: flex;
+            flex-direction: column;
+            justify-content: center;
+            align-items: center;
+            height: 100vh;
+        }
+
+        .chatbot-view {
+            display: flex;
+            flex-direction: column;
+            width: 400px;
+            height: 500px;
+            border-radius: 10px;
+            background-color: #fff;
+            overflow-y: scroll;
+            box-shadow: 0px 0px 10px #888888;
+            margin: 50px auto;
+        }
+
+        .chatbot-header {
+            background-color: #3d3d3d;
+            color: #fff;
+            font-size: 18px;
+            font-weight: bold;
+            padding: 10px;
+            border-top-left-radius: 10px;
+            border-top-right-radius: 10px;
+        }
+
+        .chatbot-messages {
+            padding: 10px;
+        }
+
+        .user-message {
+            background-color: #e6e6e6;
+            border-radius: 5px;
+            padding: 5px;
+            margin-bottom: 10px;
+            max-width: 80%;
+            float: right;
+            clear: both;
+        }
+
+        .bot-message {
+            background-color: #fff;
+            border-radius: 5px;
+            padding: 5px;
+            margin-bottom: 10px;
+            max-width: 80%;
+            float: left;
+            clear: both;
+        }
+
+        .chatbot-input {
+            display: flex;
+            padding: 10px;
+        }
+
+        .chatbot-input input[type="text"] {
+            flex: 1;
+            border: none;
+            border-radius: 5px;
+            padding: 10px;
+            font-size: 16px;
+        }
+
+        .chatbot-input input[type="submit"] {
+            background-color: #3d3d3d;
+            color: #fff;
+            border: none;
+            border-radius: 5px;
+            padding: 10px;
+            font-size: 16px;
+            cursor: pointer;
+            margin-left: 10px;
+        }
+    </style>
+</head>
+
+<body>
+    <div class="chatbot-container">
+        <div class="chatbot-view">
+            <div class="chatbot-header">ChatGPT</div>
+            <div class="chatbot-messages">
+                <div class="bot-message">Hi there! How can I help you today?</div>
+            </div>
+        </div>
+        <div class="chatbot-input">
+            <form action="/" method="POST">
+                <input type="text" placeholder="Type your message here" name="user_input">
+                <input type="submit" value="Send">
+            </form>
+        </div>
+    </div>
+
+    <script src='https://use.fontawesome.com/releases/v5.0.13/js/all.js'></script>
+    <script>
+        const msgerForm = get(".msger-inputarea");
+        const msgerInput = get(".msger-input");
+        const msgerChat = get(".msger-chat");
+        // Icons made by Freepik from www.flaticon.com
+        const BOT_IMG = "https://image.flaticon.com/icons/svg/327/327779.svg";
+        const PERSON_IMG = "https://image.flaticon.com/icons/svg/145/145867.svg";
+        const BOT_NAME = "    ChatBot";
+        const PERSON_NAME = "You";
+        msgerForm.addEventListener("submit", event => {
+            event.preventDefault();
+            const msgText = msgerInput.value;
+            if (!msgText) return;
+            appendMessage(PERSON_NAME, PERSON_IMG, "right", msgText);
+            msgerInput.value = "";
+            botResponse(msgText);
+        });
+        function appendMessage(name, img, side, text) {
+            //   Simple solution for small apps
+            const msgHTML = `
+                <div class="msg ${side}-msg">
+                <div class="msg-img" style="background-image: url(${img})"></div>
+                <div class="msg-bubble">
+                    <div class="msg-info">
+                    <div class="msg-info-name">${name}</div>
+                    <div class="msg-info-time">${formatDate(new Date())}</div>
+                    </div>
+                    <div class="msg-text">${text}</div>
+                </div>
+                </div>
+            `;
+            msgerChat.insertAdjacentHTML("beforeend", msgHTML);
+            msgerChat.scrollTop += 500;
+        }
+        function botResponse(rawText) {
+            // Bot Response
+            $.get("/get", { msg: rawText }).done(function (data) {
+                console.log(rawText);
+                console.log(data);
+                const msgText = data;
+                appendMessage(BOT_NAME, BOT_IMG, "left", msgText);
+            });
+        }
+        // Utils
+        function get(selector, root = document) {
+            return root.querySelector(selector);
+        }
+        function formatDate(date) {
+            const h = "0" + date.getHours();
+            const m = "0" + date.getMinutes();
+            return `${h.slice(-2)}:${m.slice(-2)}`;
+        }
+    </script>
+</body>
+
+</html>
