+++ web-frontend/test/webapp/botRespond.py
@@ -1,0 +1,5 @@
+import random
+
+def generate_response(user_input):
+    responses = ['Hello!', 'How can I help you?', 'Nice to meet you!']
+    return random.choice(responses)
