--- web-frontend/ui/.replit
@@ -1,2 +1,0 @@
-language = "python3"
-run = "cd webapp; python3 app.py"
