--- web-frontend/ui/pyproject.toml
@@ -1,16 +1,0 @@
-[tool.poetry]
-name = "repl_python3_build-an-ml-web-app-1"
-version = "0.1.0"
-description = ""
-authors = ["ControlBot <<>>"]
-
-[tool.poetry.dependencies]
-python = "^3.8"
-Flask = "^2.0.2"
-pandas = "^1.3.5"
-
-[tool.poetry.dev-dependencies]
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
