--- web-frontend/ui/webapp/templates/index.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <title>Chatbot</title>
-</head>
-
-<body>
-    <h1>Chatbot</h1>
-    <form action="/" method="POST">
-        <input type="text" name="user_input">
-        <input type="submit" value="Send">
-    </form>
-    {% if response %}
-    <p>{{ response }}</p>
-    {% endif %}
-</body>
-
-</html>
